fix: Made appointment ownership more strict

+++ app/Http/Controllers/Calender/AppointmentController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Calender;
 
+use App\Enums\Role;
 use App\Http\Controllers\Controller;
 use App\Models\Appointment;
 use App\Models\User;
@@ -91,6 +92,10 @@ class AppointmentController extends Controller
 
     public function store(Request $request)
     {
+        if ($request->user()->hasRole(Role::INSTRUCTOR)) {
+            return back()->with('error', 'Sie dürfen keine termine erstellen');
+        }
+
         $validated = $this->validateAppointment($request);
 
         Appointment::create([
@@ -103,6 +108,10 @@ class AppointmentController extends Controller
 
     public function update(Request $request, Appointment $appointment)
     {
+        if ($request->user()->id !== $appointment->creator()->id) {
+            return back()->with('error', 'Sie dürfen diesen termin nicht bearbeiten');
+        }
+
         $validated = $this->validateAppointment($request);
 
         $appointment->update($validated);
@@ -110,8 +119,12 @@ class AppointmentController extends Controller
         return back()->with('success', 'Termin erfolgreich aktualisiert!');
     }
 
-    public function destroy(Appointment $appointment)
+    public function destroy(Request $request, Appointment $appointment)
     {
+        if ($request->user()->id !== $appointment->creator()->id) {
+            return back()->with('error', 'Sie dürfen diesen termin nicht löschen');
+        }
+
         $appointment->delete();
 
         return back()->with('success', 'Termin erfolgreich gelöscht!');
